pur TR

+++ routes/web.php
@@ -27,31 +27,31 @@ use Illuminate\Support\Facades\Route;
 
 
 // Dashboard Routes
-Route::get('admin/',[DBController::class,'index'])->name('admin.index');
+Route::get('admin/',[DBController::class,'index'])->name('admin.index')->middleware('auth');
 
 // Store Routes
-Route::get('admin/stores',[StoreController::class,'index'])->name('admin.stores.index');
-Route::get('admin/stores/create',[StoreController::class,'create'])->name('admin.stores.create');
-Route::post('admin/stores/',[StoreController::class,'store'])->name('admin.stores.store');
-Route::get('admin/stores/edit/{store}',[StoreController::class,'edit'])->name('admin.stores.edit');
-Route::put('admin/stores/{store}',[StoreController::class,'update'])->name('admin.stores.update');
-Route::delete('admin/stores/{store}',[StoreController::class,'destroy'])->name('admin.stores.destroy');
-Route::post('admin/stores/restore/{store}',[StoreController::class,'restore'])->name('admin.stores.restore');
+Route::get('admin/stores',[StoreController::class,'index'])->name('admin.stores.index')->middleware('auth');
+Route::get('admin/stores/create',[StoreController::class,'create'])->name('admin.stores.create')->middleware('auth');
+Route::post('admin/stores/',[StoreController::class,'store'])->name('admin.stores.store')->middleware('auth');
+Route::get('admin/stores/edit/{store}',[StoreController::class,'edit'])->name('admin.stores.edit')->middleware('auth');
+Route::put('admin/stores/{store}',[StoreController::class,'update'])->name('admin.stores.update')->middleware('auth');
+Route::delete('admin/stores/{store}',[StoreController::class,'destroy'])->name('admin.stores.destroy')->middleware('auth');
+Route::post('admin/stores/restore/{store}',[StoreController::class,'restore'])->name('admin.stores.restore')->middleware('auth');
 // Route::get('admin/stores/{store}',[StoreController::class,'show'])->name('admin.stores.show');
 
 // Product Routes
-Route::get('admin/products',[ProductController::class,'index'])->name('admin.products.index');
-Route::get('admin/products/create',[ProductController::class,'create'])->name('admin.products.create');
-Route::post('admin/products/',[ProductController::class,'store'])->name('admin.products.store');
-Route::get('admin/products/edit/{product}',[ProductController::class,'edit'])->name('admin.products.edit');
-Route::put('admin/products/{product}',[ProductController::class,'update'])->name('admin.products.update');
-Route::delete('admin/products/{product}',[ProductController::class,'destroy'])->name('admin.products.destroy');
-Route::post('admin/products/restore/{product}',[ProductController::class,'restore'])->name('admin.products.restore');
+Route::get('admin/products',[ProductController::class,'index'])->name('admin.products.index')->middleware('auth');
+Route::get('admin/products/create',[ProductController::class,'create'])->name('admin.products.create')->middleware('auth');
+Route::post('admin/products/',[ProductController::class,'store'])->name('admin.products.store')->middleware('auth');
+Route::get('admin/products/edit/{product}',[ProductController::class,'edit'])->name('admin.products.edit')->middleware('auth');
+Route::put('admin/products/{product}',[ProductController::class,'update'])->name('admin.products.update')->middleware('auth');
+Route::delete('admin/products/{product}',[ProductController::class,'destroy'])->name('admin.products.destroy')->middleware('auth');
+Route::post('admin/products/restore/{product}',[ProductController::class,'restore'])->name('admin.products.restore')->middleware('auth');
 // Route::get('admin/products/{product}',[ProductController::class,'show'])->name('admin.product.show');
 
 // Purchase_transaction Route
-Route::get('admin/purchaseTransaction',[PurchaseTransactionController::class,'index'])->name('admin.PurchaseTransaction.index');
-Route::get('admin/purchaseTransaction/total',[PurchaseTransactionController::class,'total'])->name('admin.PurchaseTransaction.total');
+Route::get('admin/purchaseTransaction',[PurchaseTransactionController::class,'index'])->name('admin.PurchaseTransaction.index')->middleware('auth');
+Route::get('admin/purchaseTransaction/total',[PurchaseTransactionController::class,'total'])->name('admin.PurchaseTransaction.total')->middleware('auth');
 
 // Site Route
 Route::get('/',[SiteController::class,'index'])->name('website.index');
